requirements.txt
pandas
langchain_core
langchain_openai
langchain_anthropic
python-dotenv
scikit-optimize
opencv-python
scikit-image
colorama
scipy
regex
pydantic
numpy
opencv_python
pyqtgraph
scikit_optimize
scikit_image
watchdog
tqdm
tabulate